Intermediate commit
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/IMS/current@12458 ececa60f-9f63-0410-97c7-ea05888eb260

+++ imsApp/src/misc.h
@@ -2,10 +2,10 @@ struct ims_info
 {
     epicsMutex *cMutex;
     asynUser   *pasynUser;
-    int         S1;                                                 // switch S1
-    int         S2;
-    int         S3;
-    int         init;
+    short       S1;                                                 // switch S1
+    short       S2;
+    short       S3;
+    short       init;
     int         sword;
     long        count;
     bool        newData;
